ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
python-multipart==0.0.9
Pillow==10.3.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.7
